fix: most error codes validated and tested for
