+++ test-app/src/message-board.ts
@@ -666,3 +666,11 @@ export interface Comment<
   commentText: string;
   createdTime: string;
 }
+
+new StepFunction(stack, "obj ref", async () => {
+  const getIntegration = (): typeof func => {
+    return func;
+  };
+  const x = getIntegration();
+  await x();
+});
