+++ test/test-files/step-function.ts
@@ -219,3 +219,13 @@ new StepFunction(stack, "obj ref", async () => {
     Payload: undefined,
   });
 });
+
+// unsupported - cannot find reference to integration outside of scope.
+
+new StepFunction(stack, "obj ref", async () => {
+  const getIntegration = (): typeof func => {
+    return func;
+  };
+  const x = getIntegration();
+  await x();
+});
